requirements.txt
altgraph==0.17.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
buildozer==1.5.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.7
docutils==0.20.1
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keyboard==0.13.5
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
oauthlib==3.2.2
package-name==0.1
packaging==23.1
parso==0.8.3
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pythonnet==3.0.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sh==2.0.6
six==1.16.0
sounddevice==0.4.6
SQLAlchemy==2.0.20
stack-data==0.6.2
toga-core==0.3.1
toga-winforms==0.3.1
tornado==6.3.3
traitlets==5.9.0
travertino==0.3.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.3
Wave==0.0.2
wcwidth==0.2.6
Werkzeug==2.3.7
